update io crate
