+++ crates/io/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "key"
+name = "io"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
