+++ crates/io/src/event/mod.rs
@@ -1,0 +1,7 @@
+use crate::{KeyEvent, MovementEvent};
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum InputEvent {
+  Key(KeyEvent),
+  Movement(MovementEvent),
+}
