+++ crates/io/src/key/alphabetic/mod.rs
@@ -1,3 +1,7 @@
+mod error;
+
+pub use error::AlphabeticKeyError;
+
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -155,23 +159,6 @@ impl TryFrom<evdev::KeyCode> for AlphabeticKey {
     })
   }
 }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum AlphabeticKeyError {
-  InvalidKey(char),
-  InvalidCode(u16),
-}
-
-impl std::fmt::Display for AlphabeticKeyError {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::InvalidKey(key) => write!(f, "invalid alphabetic key: '{}'", key),
-      Self::InvalidCode(code) => write!(f, "invalid alphabetic code: '{}'", code),
-    }
-  }
-}
-
-impl std::error::Error for AlphabeticKeyError {}
 
 #[cfg(test)]
 mod tests {
