+++ crates/io/src/key/arrow/mod.rs
@@ -1,3 +1,7 @@
+mod error;
+
+pub use error::ArrowKeyError;
+
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -54,23 +58,6 @@ impl std::fmt::Display for ArrowKey {
     f.write_str(self.as_str())
   }
 }
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ArrowKeyError {
-  InvalidKey(String),
-  InvalidCode(u16),
-}
-
-impl std::fmt::Display for ArrowKeyError {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::InvalidKey(key) => write!(f, "invalid arrow key: '{}'", key),
-      Self::InvalidCode(code) => write!(f, "invalid arrow code: '{}'", code),
-    }
-  }
-}
-
-impl std::error::Error for ArrowKeyError {}
 
 #[cfg(test)]
 mod tests {
