+++ crates/io/src/key/function/mod.rs
@@ -1,3 +1,7 @@
+mod error;
+
+pub use error::FunctionKeyError;
+
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -99,25 +103,6 @@ impl std::fmt::Display for FunctionKey {
     write!(f, "{}", self.as_str())
   }
 }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum FunctionKeyError {
-  Number(u8),
-  Format,
-  Code(u16),
-}
-
-impl std::fmt::Display for FunctionKeyError {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::Number(n) => write!(f, "invalid function key: f{}", n),
-      Self::Format => write!(f, "invalid function key format"),
-      Self::Code(code) => write!(f, "invalid function key code: '{}'", code),
-    }
-  }
-}
-
-impl std::error::Error for FunctionKeyError {}
 
 #[cfg(test)]
 mod tests {
