+++ crates/io/src/key/modifier/mod.rs
@@ -1,3 +1,7 @@
+mod error;
+
+pub use error::ModifierKeyError;
+
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -100,23 +104,6 @@ impl std::fmt::Display for ModifierKey {
     f.write_str(self.as_str())
   }
 }
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ModifierKeyError {
-  InvalidKey(String),
-  InvalidCode(u16),
-}
-
-impl std::fmt::Display for ModifierKeyError {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::InvalidKey(key) => write!(f, "invalid modifier key: '{}'", key),
-      Self::InvalidCode(code) => write!(f, "invalid modifier code: '{}'", code),
-    }
-  }
-}
-
-impl std::error::Error for ModifierKeyError {}
 
 #[cfg(test)]
 mod tests {
