+++ crates/io/src/key/mouse/mod.rs
@@ -1,3 +1,7 @@
+mod error;
+
+pub use error::MouseKeyError;
+
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -58,24 +62,6 @@ impl std::fmt::Display for MouseKey {
     f.write_str(self.as_str())
   }
 }
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum MouseKeyError {
-  InvalidKey(String),
-  InvalidCode(u16),
-}
-
-impl std::fmt::Display for MouseKeyError {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::InvalidKey(s) => write!(f, "invalid mouse key: '{}'", s),
-      Self::InvalidCode(code) => write!(f, "invalid mouse key code: '{}'", code),
-    }
-  }
-}
-
-impl std::error::Error for MouseKeyError {}
-
 #[cfg(test)]
 mod tests {
   use super::*;
