+++ crates/io/src/key/numeric/mod.rs
@@ -1,3 +1,7 @@
+mod error;
+
+pub use error::NumericKeyError;
+
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -111,25 +115,6 @@ impl TryFrom<evdev::KeyCode> for NumericKey {
     })
   }
 }
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum NumericKeyError {
-  Digit(u8),
-  Key(char),
-  Code(u16),
-}
-
-impl std::fmt::Display for NumericKeyError {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::Digit(n) => write!(f, "invalid numeric key: '{}'", n),
-      Self::Key(key) => write!(f, "invalid numeric char: '{}'", key),
-      Self::Code(code) => write!(f, "invalid numeric key code: '{}'", code),
-    }
-  }
-}
-
-impl std::error::Error for NumericKeyError {}
 
 #[cfg(test)]
 mod tests {
