+++ crates/io/src/key/state.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum KeyState {
+  Pressed,
+  Released,
+  Held,
+  Moved,
+}
