+++ crates/io/src/key/system/mod.rs
@@ -1,3 +1,7 @@
+mod error;
+
+pub use error::SystemKeyError;
+
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -54,23 +58,6 @@ impl std::fmt::Display for SystemKey {
     f.write_str(self.as_str())
   }
 }
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum SystemKeyError {
-  InvalidKey(String),
-  InvalidCode(u16),
-}
-
-impl std::fmt::Display for SystemKeyError {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::InvalidKey(s) => write!(f, "invalid system key: '{}'", s),
-      Self::InvalidCode(code) => write!(f, "invalid system key code: '{}'", code),
-    }
-  }
-}
-
-impl std::error::Error for SystemKeyError {}
 
 #[cfg(test)]
 mod tests {
