+++ crates/io/src/lib.rs
@@ -1,0 +1,7 @@
+mod event;
+mod key;
+mod movement;
+
+pub use event::*;
+pub use key::*;
+pub use movement::*;
