+++ crates/io/src/movement/axis/mod.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum Axis {
+  X,
+  Y,
+}
