+++ crates/io/src/movement/error.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum MovementError {}
+
+impl std::error::Error for MovementError {}
