+++ crates/io/src/movement/event.rs
@@ -1,0 +1,7 @@
+use super::{Movement, Polarity};
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct MovementEvent {
+  kind: Movement,
+  polarity: Polarity,
+}
