+++ crates/io/src/movement/relative/mod.rs
@@ -1,0 +1,6 @@
+use super::Axis;
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct RelativeMovement {
+  axis: Axis,
+}
